Moved src files up from src/raylib, replacing the old main.cpp. Added a currently working compilation command as a comment to the main makefile, and removed the makefile that was in src/raylib. Removed helper.cpp, old main.cpp is still present, but hidden.
